Move table to a component and add full lodash
