+++ package.json
@@ -51,7 +51,7 @@
     "highcharts": "9.1.0",
     "highcharts-react-official": "3.0.0",
     "keycloak-js": "^15.0.2",
-    "lodash.truncate": "^4.4.2",
+    "lodash": "^4.17.21",
     "mobx": "6.3.2",
     "mobx-react": "7.2.0",
     "next": "^12.1.0",
